Audit: Remove duplicate tests and standardise batch test names in propagation module (#1766)
Remove AccumulateBias.ts and WeightCalculation.ts (duplicate tests consolidated
into Bias.ts, BiasConvergence.ts, and Weight.ts). Remove duplicate batch test
sections from Bias.ts and Weight.ts. Standardise 30 batch test names from
CamelCase-Hyphenated to consistent descriptive format.
