--- test/propagate/AccumulateBias.ts
@@ -1,245 +1,0 @@
-import { assertAlmostEquals } from "@std/assert";
-import { Creature, type CreatureExport } from "../../mod.ts";
-import { NeuronState } from "../../src/architecture/CreatureState.ts";
-import {
-  createBackPropagationConfig,
-} from "../../src/propagate/BackPropagation.ts";
-import { accumulateBias, adjustedBias } from "../../src/propagate/Bias.ts";
-
-Deno.test("accumulateBias - positive delta accumulates bias correctly", () => {
-  const ns = new NeuronState();
-  const config = createBackPropagationConfig({
-    generations: 0,
-    learningRate: 1,
-  });
-  accumulateBias(ns, 4, 2, 0, config);
-
-  assertAlmostEquals(ns.totalBias, 2, 0.1, JSON.stringify(ns, null, 2));
-});
-
-Deno.test("accumulateBias - large delta is clamped by adjustment limit", () => {
-  const ns = new NeuronState();
-  const config = createBackPropagationConfig({
-    generations: 0,
-    learningRate: 1,
-  });
-  accumulateBias(ns, 40, 2, 0, config);
-
-  const expected = 38;
-  assertAlmostEquals(
-    ns.totalBias,
-    expected,
-    0.1,
-    `expected: ${expected}, totalBias: ${ns.totalBias}`,
-  );
-});
-
-function makeCreature() {
-  /*
-   *  i0 i1 i2
-   *  h3=(i0 * -0.1) + (i1 * 0.2) - 0.3
-   *  o4=(h3 * 0.4) - 0.5
-   *  o5=(h3 * -0.6) + (i2 * 0.7 ) + 0.8
-   */
-  const json: CreatureExport = {
-    neurons: [
-      { type: "hidden", uuid: "hidden-3", squash: "IDENTITY", bias: 0 },
-
-      {
-        type: "output",
-        squash: "IDENTITY",
-        uuid: "output-0",
-        bias: 1,
-      },
-      {
-        type: "output",
-        squash: "IDENTITY",
-        uuid: "output-1",
-        bias: 0,
-      },
-    ],
-    synapses: [
-      /* h3=(i0 * -0.1) + (i1 * 0.2) - 0.3 */
-      { fromUUID: "input-0", toUUID: "hidden-3", weight: -1 },
-      { fromUUID: "input-1", toUUID: "hidden-3", weight: 0 },
-
-      /* o4=(h3 * 0.4) - 0.5*/
-      { fromUUID: "hidden-3", toUUID: "output-0", weight: 1 },
-
-      /* o5=(h3 * -0.6) + (i2 * 0.7 ) + 0.8*/
-      { fromUUID: "hidden-3", toUUID: "output-1", weight: 1 },
-      { fromUUID: "input-2", toUUID: "output-1", weight: 1 },
-    ],
-    input: 3,
-    output: 2,
-  };
-  const creature = Creature.fromJSON(json);
-  creature.validate();
-
-  return creature;
-}
-
-Deno.test("AccumulateBias-convergence: adjustedBias converges to target across diverse bias values", () => {
-  let config = createBackPropagationConfig({
-    generations: 0,
-    learningRate: 1,
-    batchSize: 1, // Disable mini-batching for deterministic behaviour
-  });
-
-  const creature = makeCreature();
-  const neuron = creature.neurons[3];
-
-  const biases = [100, -0.1, 0, 0.2, -0.3, 4, -5, 60, -70];
-  biases.forEach((bias) => {
-    if (bias > 50) {
-      config = createBackPropagationConfig({
-        ...config,
-        maximumBiasAdjustmentScale: 500,
-      });
-    }
-
-    creature.clearState();
-    const state = neuron.creature.state;
-    const ns = state.node(neuron.index);
-
-    const values = [
-      0.1,
-      10,
-      20,
-      30,
-      40,
-      50,
-      60,
-      70,
-      80,
-      90,
-      100,
-      0.2,
-      0.35,
-      0.5,
-      0.75,
-      1,
-      1.5,
-      2,
-      3,
-      4,
-      5,
-      6,
-      7,
-      8,
-      9,
-      10,
-      20,
-      30,
-      1,
-      40,
-      50,
-      60,
-      70,
-      80,
-      90,
-      100,
-      0.2,
-      0.35,
-      0.5,
-      0.75,
-      1,
-      1.5,
-      2,
-      3,
-      4,
-      5,
-      6,
-      7,
-      8,
-      9,
-      10,
-      20,
-      30,
-      40,
-      50,
-      60,
-      70,
-      1.5,
-      80,
-      90,
-      100,
-      0.2,
-      0.35,
-      0.5,
-      0.75,
-      1,
-      1.5,
-      2,
-      3,
-      4,
-      5,
-      6,
-      7,
-      8,
-      9,
-      10,
-      20,
-      30,
-      40,
-      50,
-      60,
-      70,
-      80,
-      90,
-      100,
-      0.2,
-      2,
-      0.35,
-      0.5,
-      0.75,
-      1,
-      1.5,
-      2,
-      3,
-      4,
-      5,
-      6,
-      7,
-      8,
-      9,
-      10,
-      20,
-      30,
-      40,
-      50,
-      60,
-      70,
-      80,
-      90,
-      100,
-      0.2,
-      0.35,
-      0.5,
-      0.75,
-    ];
-    values.forEach((targetValue) => {
-      const currentValue = targetValue - bias;
-      accumulateBias(ns, targetValue, currentValue, 0, config);
-
-      accumulateBias(
-        ns,
-        targetValue * -1,
-        targetValue * -1 - bias,
-        0,
-        config,
-      );
-    });
-
-    const aBias = adjustedBias(neuron, config);
-    const tolerance = Math.abs(bias) > 50 ? 2 : Math.abs(bias) > 3 ? 0.5 : 0.01;
-    assertAlmostEquals(
-      aBias,
-      bias,
-      tolerance,
-      `aBias: ${aBias}, bias: ${bias}, state: ${
-        JSON.stringify(ns, null, 2)
-      }, tolerance: ${tolerance}`,
-    );
-  });
-});
