+++ test/propagate/Bias.ts
@@ -15,12 +15,7 @@ import {
   assertThrows,
 } from "@std/assert";
 import { NeuronState } from "../../src/architecture/CreatureState.ts";
-import {
-  accumulateBias,
-  accumulateBiasBatch4Way,
-  accumulateBiasBatch8Way,
-  limitBias,
-} from "../../src/propagate/Bias.ts";
+import { accumulateBias, limitBias } from "../../src/propagate/Bias.ts";
 import {
   type BackPropagationConfig,
   createBackPropagationConfig,
@@ -257,71 +252,4 @@ Deno.test("limitBias - throws on non-finite targetBias", () => {
   assertThrows(() => limitBias(NaN, 0.5, config));
   assertThrows(() => limitBias(Infinity, 0.5, config));
   assertThrows(() => limitBias(-Infinity, 0, config), Error, "finite");
-});
-
-// ---------------------------------------------------------------------------
-// accumulateBiasBatch4Way
-// ---------------------------------------------------------------------------
-
-Deno.test("accumulateBiasBatch4Way - accumulates 4 neurons", () => {
-  const nsArray = Array.from({ length: 4 }, () => new NeuronState());
-  const targets = [2.0, 3.0, 4.0, 5.0];
-  const preActs = [1.0, 1.0, 1.0, 1.0];
-  const biases = [0.0, 0.0, 0.0, 0.0];
-
-  accumulateBiasBatch4Way(nsArray, targets, preActs, biases, makeConfig());
-
-  for (let i = 0; i < 4; i++) {
-    assertEquals(nsArray[i].count, 1);
-    assertAlmostEquals(
-      nsArray[i].totalBias,
-      targets[i] - preActs[i],
-      1e-10,
-    );
-  }
-});
-
-Deno.test("accumulateBiasBatch4Way - skips non-finite values per element", () => {
-  const nsArray = Array.from({ length: 4 }, () => new NeuronState());
-  const targets = [2.0, NaN, 4.0, 5.0];
-  const preActs = [1.0, 1.0, Infinity, 1.0];
-  const biases = [0.0, 0.0, 0.0, NaN];
-
-  accumulateBiasBatch4Way(nsArray, targets, preActs, biases, makeConfig());
-
-  assertEquals(nsArray[0].count, 1);
-  assertEquals(nsArray[1].count, 0); // NaN target
-  assertEquals(nsArray[2].count, 0); // Infinity preActivation
-  assertEquals(nsArray[3].count, 0); // NaN bias
-});
-
-// ---------------------------------------------------------------------------
-// accumulateBiasBatch8Way
-// ---------------------------------------------------------------------------
-
-Deno.test("accumulateBiasBatch8Way - accumulates 8 neurons", () => {
-  const nsArray = Array.from({ length: 8 }, () => new NeuronState());
-  const targets = [1, 2, 3, 4, 5, 6, 7, 8];
-  const preActs = [0, 0, 0, 0, 0, 0, 0, 0];
-  const biases = [0, 0, 0, 0, 0, 0, 0, 0];
-
-  accumulateBiasBatch8Way(nsArray, targets, preActs, biases, makeConfig());
-
-  for (let i = 0; i < 8; i++) {
-    assertEquals(nsArray[i].count, 1);
-    assertAlmostEquals(nsArray[i].totalBias, targets[i], 1e-10);
-  }
-});
-
-Deno.test("accumulateBiasBatch8Way - skips non-finite values selectively", () => {
-  const nsArray = Array.from({ length: 8 }, () => new NeuronState());
-  const targets = [1, NaN, 3, 4, 5, 6, 7, 8];
-  const preActs = [0, 0, 0, 0, 0, 0, 0, 0];
-  const biases = [0, 0, 0, 0, 0, 0, 0, 0];
-
-  accumulateBiasBatch8Way(nsArray, targets, preActs, biases, makeConfig());
-
-  assertEquals(nsArray[0].count, 1);
-  assertEquals(nsArray[1].count, 0);
-  assertEquals(nsArray[2].count, 1);
 });
